feat(rapport-html): Génération d'un rapport au format HTML
