+++ cli-core/template/page.html
@@ -1,0 +1,163 @@
+<!doctype html>
+<html lang="en">
+
+<head>
+    <!-- Required meta tags -->
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+
+    <!-- Bootstrap CSS -->
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-wEmeIV1mKuiNpC+IOBjI7aAzPcEZeedi5yW5f2yOq55WWLwNGmvvx4Um1vskeMj0" crossorigin="anonymous">
+
+    <style type="text/css"> 
+        .grade {
+            display: inline-block;
+            width: 24px;
+            height: 24px;
+            text-align: center;
+            border-radius: 50%;
+            background-color: #f00;
+            font-size: 14px;
+            font-weight: 650;
+            line-height: 24px;
+        }
+
+        .grade.big-grade {
+            width: 42px;
+            height: 42px;
+            font-size: 28px;
+            line-height: 42px;
+        }
+
+        .grade.A {
+            background-color: #349A47;
+            color: #fff;
+        }
+    
+        .grade.B {
+            background-color: #51B84B;
+            color: #fff;
+        }
+    
+        .grade.C {
+            background-color: #CADB2A;
+            color: #fff;
+        }
+    
+        .grade.D {
+            background-color: #F6EB15;
+        }
+    
+        .grade.E {
+            background-color: #FECD06;
+        }
+    
+        .grade.F {
+            background-color: #F99839;
+            color: #fff;
+        }
+    
+        .grade.G {
+            background-color: #ED2124;
+            color: #fff;
+        }
+        </style>
+
+    <title>GreenIT-Analysis report</title>
+</head>
+
+<body>
+    <nav class="navbar navbar-success bg-success text-white mb-4">
+        <div class="container-fluid">
+            <span class="navbar-brand mb-0 h1" style="font-size: 1rem;" thjs:utext="${header}">GreenIT-Analysis report > <a class="text-white" href="http://10.132.7.169:81/conso/supervision/supervision/suivi-traitements">RPT | Données patrimoniales réseau</a></span>
+            <div class="d-flex flex-row-reverse">
+                <div thjs:text="${date}">05/05/2021 10:03:12</div>
+            </div>
+        </div>
+    </nav>
+    <div class="container">
+        <div class="container my-5">
+            <div class="row text-center">
+                <div class="col col-md-4">
+                    <div class="card mb-4 border-success" style="max-width: 18rem;">
+                        <div class="card-header bg-success text-white">Ecoindex</div>
+                        <div class="card-body text-dark">
+                            <h1 class="card-title" style="height: 45px;" thjs:utext="${bigEcoIndex}">51 <span class="grade big-grade C">C</span></h1>
+                        </div>
+                    </div>
+                </div>
+
+                <div class="col col-md-4">
+                    <div class="card mb-4 border-success" style="max-width: 18rem;">
+                        <div class="card-header bg-success text-white">Eau (cl)</div>
+                        <div class="card-body text-dark">
+                            <h1 class="card-title" thjs:text="${waterConsumption}">3.24</h1>
+                        </div>
+                    </div>
+                </div>
+
+                <div class="col col-md-4">
+                    <div class="card mb-4 border-success" style="max-width: 18rem;">
+                        <div class="card-header bg-success text-white">GES (gCO2e)</div>
+                        <div class="card-body text-dark">
+                            <h1 class="card-title" thjs:text="${greenhouseGasesEmission}">2.16</h1>
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <div class="row text-center">
+                <div class="col col-md-4">
+                    <div class="card mb-4 border-success" style="max-width: 18rem;">
+                        <div class="card-header bg-success text-white">Requêtes</div>
+                        <div class="card-body text-dark">
+                            <h1 class="card-title" thjs:text="${nbRequest}">18</h1>
+                        </div>
+                    </div>
+                </div>
+
+                <div class="col col-md-4">
+                    <div class="card mb-4 border-success" style="max-width: 18rem;">
+                        <div class="card-header bg-success text-white">Taille de la page (Ko)</div>
+                        <div class="card-body text-dark">
+                            <h1 class="card-title" thjs:text="${pageSize}">3791 (3787)</h1>
+                        </div>
+                    </div>
+                </div>
+
+                <div class="col col-md-4">
+                    <div class="card mb-4 border-success" style="max-width: 18rem;">
+                        <div class="card-header bg-success text-white">Taille du DOM</div>
+                        <div class="card-body text-dark">
+                            <h1 class="card-title" thjs:text="${domSize}">951</h1>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+        <div class="container my-3">
+            <h5 class="mb-3">Bonnes pratiques</h5>
+            <div class="row">
+                <div class="col col-md-12">
+                    <table class="table table-sm">
+                        <thead class="table-light">
+                            <th scope="col">Règle</th>
+                            <th scope="col">Résultat</th>
+                            <th scope="col">Note</th>
+                        </thead>
+                        <tbody thjs:each="bestPractice: ${bestPractices}">
+                            <tr>
+                                <td thjs:text=${bestPractice.name}>Ajouter des expires ou cache-control headers (>= 95%)</td>
+                                <td thjs:text=${bestPractice.comment}>0% ressources cachées</td>
+                                <td><img thjs:src="${bestPractice.iconGrade}" src="icons/A.png"></td>
+                            </tr>
+                        </tbody>
+                    </table>
+                </div>
+            </div>
+        </div>
+    </div>
+</body>
+
+</html>
