+++ locales/fr.json
@@ -33,7 +33,7 @@
 	"rule_HttpRequests": "Limiter le nombre de requêtes HTTP (< 27)",
 	"rule_HttpRequests_DetailDescription": "Le temps de téléchargement d’une page côté client est directement corrélé au nombre et à la taille des fichiers que le navigateur doit télécharger. Pour chaque fichier, le navigateur envoie un HTTP GET au serveur. Il attend la réponse, puis télécharge la ressource dès qu'elle est disponible. Selon le type de serveur Web que vous utilisez, plus le nombre de demandes par page est élevé, moins le serveur peut gérer de pages. La réduction du nombre de requêtes par page est essentielle pour réduire le nombre de serveurs HTTP nécessaires à l'exécution du site Web et, partant, son impact sur l'environnement.",
 	"rule_HttpRequests_Comment": "%s requête(s) HTTP ",
-	"rule_ImageDownloadedNotDisplayed": "Ne télécharger pas des images inutilement",
+	"rule_ImageDownloadedNotDisplayed": "Ne pas télécharger des images inutilement",
 	"rule_ImageDownloadedNotDisplayed_DetailDescription": "Télécharger des images qui ne seront pas nécessairement visibles consomme inutilement des ressources. Il s'agit par exemple d'images qui sont affichées uniquement suite à une action utilisateur.",
 	"rule_ImageDownloadedNotDisplayed_Comment": "%s image(s) téléchargée(s) mais non affichée(s) dans la page ",
 	"rule_ImageDownloadedNotDisplayed_DetailComment": "%s de taille %s n'est pas affichée",
@@ -41,7 +41,7 @@
 	"rule_JsValidate_DetailDescription": "JSLint est un outil de contrôle de qualité du code qui vérifie que la syntaxe JavaScript utilisée sera comprise par tous les navigateurs. Le code produit est donc conforme aux règles de codage qui permettent aux interpréteurs d’exécuter le code rapidement et facilement. Le processeur est donc utilisé pour un temps plus court.",
 	"rule_JsValidate_DefaultComment": "Javascript validé",
 	"rule_JsValidate_Comment": "%s erreur(s) javascript",
-	"rule_MaxCookiesLength": "Taille maximum des cookies par domaine(<512 Octets)",
+	"rule_MaxCookiesLength": "Taille maximum des cookies par domaine(< 512 Octets)",
 	"rule_MaxCookiesLength_DetailDescription": "La longueur du cookie doit être réduite car il est envoyé à chaque requête.",
 	"rule_MaxCookiesLength_DefaultComment": "Pas de cookies",
 	"rule_MaxCookiesLength_Comment": "Taille maximum = %s Octets ",
