+++ package.json
@@ -23,6 +23,7 @@
     "puppeteer": "^5.5.0",
     "puppeteer-har": "^1.1.2",
     "sitemapper": "^3.1.12",
+    "thymeleaf": "^0.20.5",
     "yaml": "^1.10.0",
     "yargs": "^16.2.0"
   },
